+++ Changelog.md
@@ -1,4 +1,6 @@
 Change log:
+* 2023-05-02 - U35.1.1 - 4 new + 1 newly available
+  * Anniversary Celebration Goat, Rust Corgi, Homesteader's Draught Horse, Dappled Grey Draught Horse, Black Draught Horse
 * 2023-03-23 - U35 - 2 new + 2 newly available
   * Rat King, Sewer Rat
   * Twilight Fire-fly, Amber Fire-fly